Fix Clang warnings `-Wimplicit-int-conversion`

+++ src/Archive/HuffLZ.cpp
@@ -195,9 +195,9 @@ namespace OP2Utility::Archive
 
 
 	// Performs a tree lookup guided by the bitstream to find the next code
-	int HuffLZ::GetNextCode()
+	unsigned short HuffLZ::GetNextCode()
 	{
-		int nodeIndex;
+		unsigned short nodeIndex;
 		bool bBit;
 
 		// Use bitstream to find a terminal node
